# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "description": "统一错误响应。\n\n``request_id`` 是把用户看到的报错和服务端日志连起来的线索：\n用户把它报过来，照着去日志里搜就能找到这次请求的完整经过。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "title": "Problems",
            "type": "array"
          },
          "request_id": {
            "default": "",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "NotificationOut": {
        "description": "一条通知。\n\n``title`` / ``body`` 是产生时就写好的文本，不是模板参数——\n通知是「当时告诉过你这件事」的记录，事后对象改名不该改写它。",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mandatory": {
            "title": "Mandatory",
            "type": "boolean"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetType"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body",
          "workspace_id",
          "target_type",
          "target_id",
          "mandatory",
          "created_at",
          "read_at"
        ],
        "title": "NotificationOut",
        "type": "object"
      },
      "NotificationType": {
        "description": "通知的种类。前端据此选图标，V1 加偏好设置时也按它分组。\n\n和 :class:`ActivityAction` 刻意分开：**不是一一对应的**。\n一次操作可能产生一条活动和零条通知（比如自己改自己的空间设置），\n也可能产生一条活动和多条通知。硬凑成一个枚举会逼着两边互相迁就。",
        "enum": [
          "workspace_invited",
          "member_removed",
          "role_changed",
          "ownership_received",
          "run_succeeded",
          "run_failed",
          "run_submit_failed"
        ],
        "title": "NotificationType",
        "type": "string"
      },
...
      "PageOut_NotificationOut_": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "has_more"
        ],
        "title": "PageOut[NotificationOut]",
        "type": "object"
      },
...
      "TargetType": {
        "description": "一条记录指向的对象类型。前端据此决定链接跳到哪里。\n\n活动和通知共用这一套：两者指向的都是同一批领域对象，\n分成两个枚举只会让前端的跳转逻辑写两遍。",
        "enum": [
          "workspace",
          "member",
          "project",
          "project_version",
          "run"
        ],
        "title": "TargetType",
        "type": "string"
      },
      "UnreadCountOut": {
        "properties": {
          "unread": {
            "title": "Unread",
            "type": "integer"
          }
        },
        "required": [
          "unread"
        ],
        "title": "UnreadCountOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "面向 USTC 107 算力平台的协作式计算工作空间。\n\n当前迁移实现支持本地 Mock 执行闭环；真实 Worker、Git / Shared FS 和 Slurm 链路尚未完成验证。",
    "title": "107 Workspace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "description": "只看未读",
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "只看未读",
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "description": "页码，从 1 开始",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "页码，从 1 开始",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "每页条数",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "每页条数",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut_NotificationOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "请求不合法"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象可见，但当前角色无权执行该操作"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象不存在，或当前用户没有发现权限"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "与现有状态冲突，例如重名或对象不可修改"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "参数校验或提交前检查未通过，problems 列出全部原因"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "底层调度系统返回错误"
          }
        },
        "summary": "我的通知",
        "tags": [
          "notification"
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "operationId": "mark_all_read_api_v1_notifications_read_all_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "请求不合法"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象可见，但当前角色无权执行该操作"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象不存在，或当前用户没有发现权限"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "与现有状态冲突，例如重名或对象不可修改"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "参数校验或提交前检查未通过，problems 列出全部原因"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "底层调度系统返回错误"
          }
        },
        "summary": "全部标记为已读",
        "tags": [
          "notification"
        ]
      }
    },
    "/api/v1/notifications/unread-count": {
      "get": {
        "operationId": "unread_count_api_v1_notifications_unread_count_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "请求不合法"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象可见，但当前角色无权执行该操作"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象不存在，或当前用户没有发现权限"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "与现有状态冲突，例如重名或对象不可修改"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "参数校验或提交前检查未通过，problems 列出全部原因"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "底层调度系统返回错误"
          }
        },
        "summary": "未读数",
        "tags": [
          "notification"
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_read_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "title": "Notification Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "请求不合法"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象可见，但当前角色无权执行该操作"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "对象不存在，或当前用户没有发现权限"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "与现有状态冲突，例如重名或对象不可修改"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "参数校验或提交前检查未通过，problems 列出全部原因"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "底层调度系统返回错误"
          }
        },
        "summary": "标记一条为已读",
        "tags": [
          "notification"
        ]
      }
    },
...
  }
}